requirements.txt
streamlit
openai
python-dotenv
scikit-learn
PyPDF2
python-docx
tiktoken
